# docs/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp AI Agent Platform API",
    "description": "Production-grade SaaS WhatsApp AI Agent Platform with Multi-Agent & Graph RAG architecture",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@whatsappagent.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/users/profile": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user profile",
        "description": "Retrieve the authenticated user's profile",
        "operationId": "getProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/UserProfile" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "description": "Update the authenticated user's profile",
        "operationId": "updateProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/UserProfile" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/account": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user account",
        "description": "Permanently delete the authenticated user's account",
        "operationId": "deleteAccount",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Account deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/preferences": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user preferences",
        "description": "Retrieve the authenticated user's preferences",
        "operationId": "getPreferences",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Preferences retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "preferences": { "$ref": "#/components/schemas/UserPreferences" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user preferences",
        "description": "Update the authenticated user's preferences",
        "operationId": "updatePreferences",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "preferences": { "$ref": "#/components/schemas/UserPreferences" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/avatar": {
      "post": {
        "tags": ["Users"],
        "summary": "Upload avatar",
        "description": "Upload a new avatar image",
        "operationId": "uploadAvatar",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["avatarUrl"],
                "properties": {
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL of the avatar image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "avatarUrl": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": { "type": "string" }
            }
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "plan": { "type": "string", "enum": ["FREE", "PRO", "ENTERPRISE"] },
          "timezone": { "type": "string" },
          "avatarUrl": { "type": "string", "format": "uri" },
          "isEmailVerified": { "type": "boolean" },
          "isPhoneVerified": { "type": "boolean" },
          "lastActiveAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserPreferences": {
        "type": "object",
        "properties": {
          "language": { "type": "string", "example": "en" },
          "notifications": { "type": "boolean" },
          "timezone": { "type": "string", "example": "UTC" },
          "tone": { "type": "string", "enum": ["formal", "friendly", "casual"] }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string", "minLength": 2 },
          "phone": { "type": "string" },
          "timezone": { "type": "string" },
          "avatarUrl": { "type": "string", "format": "uri" }
        }
      },
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "language": { "type": "string" },
          "notifications": { "type": "boolean" },
          "timezone": { "type": "string" },
          "tone": { "type": "string", "enum": ["formal", "friendly", "casual"] }
        }
      },
...
    }
  }
}